+++ src/test/kotlin/lotto/service/LottoIssueServiceTest.kt
@@ -9,10 +9,10 @@ import org.junit.jupiter.api.Test
 class LottoIssueServiceTest {
     @Test
     fun `로또 구입 금액을 입력하면 구입 금액에 해당하는 로또를 발급한다`() {
-        val purchaseAmount = 14000
+        val purchaseAmount = 14000L
 
         val actual = LottoIssueService.issue(purchaseAmount)
 
-        assertThat(actual).hasSize(purchaseAmount / Lotto.PRICE)
+        assertThat(actual).hasSize((purchaseAmount / Lotto.PRICE).toInt())
     }
 }
